backend/requirements.txt
alembic==1.3.1
attrs==19.3.0
Click==7.0
coverage==4.5.4
ecdsa==0.14.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.0
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.5
pytest==5.3.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-jose==3.1.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
